[fix] Serializer周りを整理
