+++ network/src/commonMain/kotlin/com/wsr/knist/network/NetworkSerializer.kt
@@ -1,5 +1,3 @@
-@file:OptIn(kotlinx.serialization.ExperimentalSerializationApi::class)
-
 package com.wsr.knist.network
 
 import com.wsr.knist.network.converter.Converter
@@ -125,334 +123,93 @@ import com.wsr.knist.network.process.reshape.reshape.ReshapeD3ToD2
 import com.wsr.knist.network.process.reshape.token.TokenEmbeddingD1ToD2
 import kotlin.jvm.JvmName
 import kotlin.reflect.KClass
-import kotlinx.serialization.KSerializer
-import kotlinx.serialization.PolymorphicSerializer
-import kotlinx.serialization.builtins.ListSerializer
+import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.cbor.Cbor
-import kotlinx.serialization.descriptors.SerialDescriptor
-import kotlinx.serialization.descriptors.buildClassSerialDescriptor
-import kotlinx.serialization.encoding.CompositeDecoder
-import kotlinx.serialization.encoding.Decoder
-import kotlinx.serialization.encoding.Encoder
-import kotlinx.serialization.encoding.decodeStructure
-import kotlinx.serialization.encoding.encodeStructure
 import kotlinx.serialization.json.Json
-import kotlinx.serialization.json.okio.decodeFromBufferedSource
-import kotlinx.serialization.json.okio.encodeToBufferedSink
 import kotlinx.serialization.modules.SerializersModule
 import kotlinx.serialization.modules.plus
 import kotlinx.serialization.modules.polymorphic
 import kotlinx.serialization.modules.subclass
-import okio.BufferedSink
-import okio.BufferedSource
 
-@Suppress("UNCHECKED_CAST")
-private object GraphSourceSerializer : KSerializer<Graph.Source<*>> {
-    private val idSerializer = GraphId.serializer()
-    private val converterSerializer = PolymorphicSerializer(Converter::class) as KSerializer<Converter<Any?>>
-
-    override val descriptor: SerialDescriptor =
-        buildClassSerialDescriptor("com.wsr.knist.network.Graph.Source") {
-            element("id", idSerializer.descriptor)
-            element("converter", converterSerializer.descriptor)
+object NetworkSerializer {
+    @JvmName("registerProcess")
+    inline fun <reified T : Process> register(clazz: KClass<T>) {
+        val module = SerializersModule {
+            polymorphic(Process::class) {
+                subclass(clazz)
+            }
         }
-
-    override fun serialize(encoder: Encoder, value: Graph.Source<*>) {
-        encoder.encodeStructure(descriptor) {
-            encodeSerializableElement(descriptor, 0, idSerializer, value.id)
-            encodeSerializableElement(descriptor, 1, converterSerializer, value.converter as Converter<Any?>)
-        }
+        networkSerializerModules.add(module)
     }
 
-    override fun deserialize(decoder: Decoder): Graph.Source<*> = decoder.decodeStructure(descriptor) {
-        if (decodeSequentially()) {
-            Graph.Source(
-                id = decodeSerializableElement(descriptor, 0, idSerializer),
-                converter = decodeSerializableElement(descriptor, 1, converterSerializer),
-            )
-        } else {
-            var id: GraphId? = null
-            var converter: Converter<Any?>? = null
-            while (true) {
-                when (val index = decodeElementIndex(descriptor)) {
-                    0 -> id = decodeSerializableElement(descriptor, 0, idSerializer)
-                    1 -> converter = decodeSerializableElement(descriptor, 1, converterSerializer)
-                    CompositeDecoder.DECODE_DONE -> break
-                    else -> error("Unexpected index: $index")
-                }
+    @JvmName("registerOptimizer")
+    inline fun <reified T : Optimizer> register(clazz: KClass<T>) {
+        val module = SerializersModule {
+            polymorphic(Optimizer::class) {
+                subclass(clazz)
             }
-            Graph.Source(id = checkNotNull(id), converter = checkNotNull(converter))
         }
-    }
-}
-
-@Suppress("UNCHECKED_CAST")
-private object GraphSinkSerializer : KSerializer<Graph.Sink<*>> {
-    private val idSerializer = GraphId.serializer()
-    private val converterSerializer = PolymorphicSerializer(Converter::class) as KSerializer<Converter<Any?>>
-    private val outputSerializer = PolymorphicSerializer(Output::class)
-
-    override val descriptor: SerialDescriptor =
-        buildClassSerialDescriptor("com.wsr.knist.network.Graph.Sink") {
-            element("id", idSerializer.descriptor)
-            element("from", idSerializer.descriptor)
-            element("output", outputSerializer.descriptor)
-            element("converter", converterSerializer.descriptor)
-        }
-
-    override fun serialize(encoder: Encoder, value: Graph.Sink<*>) {
-        encoder.encodeStructure(descriptor) {
-            encodeSerializableElement(descriptor, 0, idSerializer, value.id)
-            encodeSerializableElement(descriptor, 1, idSerializer, value.from)
-            encodeSerializableElement(descriptor, 2, outputSerializer, value.output)
-            encodeSerializableElement(descriptor, 3, converterSerializer, value.converter as Converter<Any?>)
-        }
+        networkSerializerModules.add(module)
     }
 
-    override fun deserialize(decoder: Decoder): Graph.Sink<*> = decoder.decodeStructure(descriptor) {
-        if (decodeSequentially()) {
-            Graph.Sink(
-                id = decodeSerializableElement(descriptor, 0, idSerializer),
-                from = decodeSerializableElement(descriptor, 1, idSerializer),
-                output = decodeSerializableElement(descriptor, 2, outputSerializer),
-                converter = decodeSerializableElement(descriptor, 3, converterSerializer),
-            )
-        } else {
-            var id: GraphId? = null
-            var from: GraphId? = null
-            var output: Output? = null
-            var converter: Converter<Any?>? = null
-            while (true) {
-                when (val index = decodeElementIndex(descriptor)) {
-                    0 -> id = decodeSerializableElement(descriptor, 0, idSerializer)
-                    1 -> from = decodeSerializableElement(descriptor, 1, idSerializer)
-                    2 -> output = decodeSerializableElement(descriptor, 2, outputSerializer)
-                    3 -> converter = decodeSerializableElement(descriptor, 3, converterSerializer)
-                    CompositeDecoder.DECODE_DONE -> break
-                    else -> error("Unexpected index: $index")
-                }
+    @JvmName("registerWeightInitializer")
+    inline fun <reified T : WeightInitializer> register(clazz: KClass<T>) {
+        val module = SerializersModule {
+            polymorphic(WeightInitializer::class) {
+                subclass(clazz)
             }
-            Graph.Sink(
-                id = checkNotNull(id),
-                from = checkNotNull(from),
-                output = checkNotNull(output),
-                converter = checkNotNull(converter),
-            )
         }
-    }
-}
-
-class GraphNetworkSerializer<T : GraphNetwork<T>>(
-    private val factory: (
-        sources: List<Graph.Source<*>>,
-        graph: List<Graph.Node>,
-        sinks: List<Graph.Sink<*>>,
-        optimizer: Optimizer,
-        initializer: WeightInitializer,
-    ) -> T,
-) : KSerializer<T> {
-    private val sourceListSerializer = ListSerializer(GraphSourceSerializer)
-    private val nodeSerializer = ListSerializer(PolymorphicSerializer(Graph.Node::class))
-    private val sinkListSerializer = ListSerializer(GraphSinkSerializer)
-    private val optimizerSerializer = PolymorphicSerializer(Optimizer::class)
-    private val initializerSerializer = PolymorphicSerializer(WeightInitializer::class)
-
-    override val descriptor: SerialDescriptor =
-        buildClassSerialDescriptor("com.wsr.knist.network.GraphNetwork") {
-            element("sources", sourceListSerializer.descriptor)
-            element("graph", nodeSerializer.descriptor)
-            element("sinks", sinkListSerializer.descriptor)
-            element("optimizer", optimizerSerializer.descriptor)
-            element("initializer", initializerSerializer.descriptor)
-        }
-
-    override fun serialize(encoder: Encoder, value: T) {
-        encoder.encodeStructure(descriptor) {
-            encodeSerializableElement(descriptor, 0, sourceListSerializer, value.sources)
-            encodeSerializableElement(descriptor, 1, nodeSerializer, value.graph)
-            encodeSerializableElement(descriptor, 2, sinkListSerializer, value.sinks)
-            encodeSerializableElement(descriptor, 3, optimizerSerializer, value.optimizer)
-            encodeSerializableElement(descriptor, 4, initializerSerializer, value.initializer)
-        }
+        networkSerializerModules.add(module)
     }
 
-    override fun deserialize(decoder: Decoder): T = decoder.decodeStructure(descriptor) {
-        if (decodeSequentially()) {
-            factory(
-                decodeSerializableElement(descriptor, 0, sourceListSerializer),
-                decodeSerializableElement(descriptor, 1, nodeSerializer),
-                decodeSerializableElement(descriptor, 2, sinkListSerializer),
-                decodeSerializableElement(descriptor, 3, optimizerSerializer),
-                decodeSerializableElement(descriptor, 4, initializerSerializer),
-            )
-        } else {
-            var sources: List<Graph.Source<*>>? = null
-            var graph: List<Graph.Node>? = null
-            var sinks: List<Graph.Sink<*>>? = null
-            var optimizer: Optimizer? = null
-            var initializer: WeightInitializer? = null
-            while (true) {
-                when (val index = decodeElementIndex(descriptor)) {
-                    0 -> sources = decodeSerializableElement(descriptor, 0, sourceListSerializer)
-                    1 -> graph = decodeSerializableElement(descriptor, 1, nodeSerializer)
-                    2 -> sinks = decodeSerializableElement(descriptor, 2, sinkListSerializer)
-                    3 -> optimizer = decodeSerializableElement(descriptor, 3, optimizerSerializer)
-                    4 -> initializer = decodeSerializableElement(descriptor, 4, initializerSerializer)
-                    CompositeDecoder.DECODE_DONE -> break
-                    else -> error("Unexpected index: $index")
-                }
+    @JvmName("registerOptimizerD1")
+    inline fun <reified T : Optimizer.D1> register(clazz: KClass<T>) {
+        val module = SerializersModule {
+            polymorphic(Optimizer.D1::class) {
+                subclass(clazz)
             }
-            factory(
-                checkNotNull(sources),
-                checkNotNull(graph),
-                checkNotNull(sinks),
-                checkNotNull(optimizer),
-                checkNotNull(initializer),
-            )
         }
+        networkSerializerModules.add(module)
     }
-}
 
-@Suppress("UNCHECKED_CAST")
-class NetworkSerializer<I, O> :
-    KSerializer<Network<I, O>> by GraphNetworkSerializer(
-        factory = { sources, graph, sinks, optimizer, initializer ->
-            check(sources.size == 1) { "invalid Network format. sources.size=${sources.size}, expected 1" }
-            check(sinks.size == 1) { "invalid Network format. sinks.size=${sinks.size}, expected 1" }
-            Network(
-                source = sources[0] as Graph.Source<I>,
-                graph = graph,
-                sink = sinks[0] as Graph.Sink<O>,
-                optimizer = optimizer,
-                initializer = initializer,
-            )
-        },
-    ) {
-    companion object {
-        fun <I, O> encodeToString(value: Network<I, O>) = json.encodeToString(
-            serializer = NetworkSerializer(),
-            value = value,
-        )
-
-        fun <I, O> encodeToBufferedSink(value: Network<I, O>, sink: BufferedSink) = json.encodeToBufferedSink(
-            serializer = NetworkSerializer(),
-            value = value,
-            sink = sink,
-        )
-
-        fun <I, O> decodeFromString(value: String) = json.decodeFromString<Network<I, O>>(
-            deserializer = NetworkSerializer(),
-            string = value,
-        )
-
-        fun <I, O> decodeFromBufferedSource(source: BufferedSource) = json.decodeFromBufferedSource<Network<I, O>>(
-            deserializer = NetworkSerializer(),
-            source = source,
-        )
-
-        fun <I, O> encodeToCbor(value: Network<I, O>): ByteArray = cbor.encodeToByteArray(NetworkSerializer(), value)
-
-        fun <I, O> encodeToCborSink(value: Network<I, O>, sink: BufferedSink) {
-            sink.write(cbor.encodeToByteArray(NetworkSerializer<I, O>(), value))
+    @JvmName("registerOptimizerD2")
+    inline fun <reified T : Optimizer.D2> register(clazz: KClass<T>) {
+        val module = SerializersModule {
+            polymorphic(Optimizer.D2::class) {
+                subclass(clazz)
+            }
         }
+        networkSerializerModules.add(module)
+    }
 
-        fun <I, O> decodeFromCbor(bytes: ByteArray): Network<I, O> =
-            cbor.decodeFromByteArray(NetworkSerializer(), bytes)
-
-        fun <I, O> decodeFromCborSource(source: BufferedSource): Network<I, O> =
-            cbor.decodeFromByteArray(NetworkSerializer(), source.readByteArray())
-
-        val modules = mutableListOf(buildInSerializersModule)
-
-        @JvmName("registerProcess")
-        inline fun <reified T : Process> register(clazz: KClass<T>) {
-            val module = SerializersModule {
-                polymorphic(Process::class) {
-                    subclass(clazz)
-                }
+    @JvmName("registerOptimizerD3")
+    inline fun <reified T : Optimizer.D3> register(clazz: KClass<T>) {
+        val module = SerializersModule {
+            polymorphic(Optimizer.D3::class) {
+                subclass(clazz)
             }
-            modules.add(module)
         }
+        networkSerializerModules.add(module)
+    }
 
-        @JvmName("registerOptimizer")
-        inline fun <reified T : Optimizer> register(clazz: KClass<T>) {
-            val module = SerializersModule {
-                polymorphic(Optimizer::class) {
-                    subclass(clazz)
-                }
+    @JvmName("registerAttentionBiasD2")
+    inline fun <reified T : AttentionBiasD2> register(clazz: KClass<T>) {
+        val module = SerializersModule {
+            polymorphic(AttentionBiasD2::class) {
+                subclass(clazz)
             }
-            modules.add(module)
         }
+        networkSerializerModules.add(module)
+    }
 
-        @JvmName("registerWeightInitializer")
-        inline fun <reified T : WeightInitializer> register(clazz: KClass<T>) {
-            val module = SerializersModule {
-                polymorphic(WeightInitializer::class) {
-                    subclass(clazz)
-                }
+    @JvmName("registerConverter")
+    inline fun <reified T : Converter<*>> register(clazz: KClass<T>) {
+        val module = SerializersModule {
+            polymorphic(Converter::class) {
+                subclass(clazz)
             }
-            modules.add(module)
         }
-
-        @JvmName("registerOptimizerD1")
-        inline fun <reified T : Optimizer.D1> register(clazz: KClass<T>) {
-            val module = SerializersModule {
-                polymorphic(Optimizer.D1::class) {
-                    subclass(clazz)
-                }
-            }
-            modules.add(module)
-        }
-
-        @JvmName("registerOptimizerD2")
-        inline fun <reified T : Optimizer.D2> register(clazz: KClass<T>) {
-            val module = SerializersModule {
-                polymorphic(Optimizer.D2::class) {
-                    subclass(clazz)
-                }
-            }
-            modules.add(module)
-        }
-
-        @JvmName("registerOptimizerD3")
-        inline fun <reified T : Optimizer.D3> register(clazz: KClass<T>) {
-            val module = SerializersModule {
-                polymorphic(Optimizer.D3::class) {
-                    subclass(clazz)
-                }
-            }
-            modules.add(module)
-        }
-
-        @JvmName("registerAttentionBiasD2")
-        inline fun <reified T : AttentionBiasD2> register(clazz: KClass<T>) {
-            val module = SerializersModule {
-                polymorphic(AttentionBiasD2::class) {
-                    subclass(clazz)
-                }
-            }
-            modules.add(module)
-        }
-
-        @JvmName("registerConverter")
-        inline fun <reified T : Converter<*>> register(clazz: KClass<T>) {
-            val module = SerializersModule {
-                polymorphic(Converter::class) {
-                    subclass(clazz)
-                }
-            }
-            modules.add(module)
-        }
-
-        private val json
-            get() = Json {
-                serializersModule = modules.reduce { acc, module -> acc + module }
-            }
-
-        private val cbor
-            get() = Cbor {
-                serializersModule = modules.reduce { acc, module -> acc + module }
-            }
+        networkSerializerModules.add(module)
     }
 }
 
@@ -659,3 +416,17 @@ private val buildInSerializersModule = SerializersModule {
         subclass(Graph.Node.Observe::class)
     }
 }
+
+@PublishedApi
+internal val networkSerializerModules = mutableListOf(buildInSerializersModule)
+
+internal val networkSerializerJson
+    get() = Json {
+        serializersModule = networkSerializerModules.reduce { acc, module -> acc + module }
+    }
+
+@OptIn(ExperimentalSerializationApi::class)
+internal val networkSerializerCbor
+    get() = Cbor {
+        serializersModule = networkSerializerModules.reduce { acc, module -> acc + module }
+    }
